Updated version of Conditional functionality

+++ SimdSse2Conditional_amd64.go
@@ -8,26 +8,26 @@ import (
 )
 
 //go:noescape
-func _SimdSse2ConditionalCount8u(src unsafe.Pointer, stride, width, height uint64, value uint8, compareType uint8, count unsafe.Pointer)
+func _SimdSse2ConditionalCount8u(src unsafe.Pointer, stride, width, height uint64, value uint64/*uint8*/, compareType uint64/*uint8*/, count unsafe.Pointer)
 
 //go:noescape
-func _SimdSse2ConditionalCount16i(src unsafe.Pointer, stride, width, height uint64, value int16, compareType uint8, count unsafe.Pointer)
+func _SimdSse2ConditionalCount16i(src unsafe.Pointer, stride, width, height uint64, value uint64/*int16*/, compareType uint64/*uint8*/, count unsafe.Pointer)
 
 //go:noescape
-func _SimdSse2ConditionalSum(src unsafe.Pointer, srcStride, width, height uint64, mask unsafe.Pointer, maskStride uint64, value uint8, compareType uint8, sum unsafe.Pointer)
+func _SimdSse2ConditionalSum(src unsafe.Pointer, srcStride, width, height uint64, mask unsafe.Pointer, maskStride uint64, value uint64/*uint8*/, compareType uint64/*uint8*/, sum unsafe.Pointer)
 
 //go:noescape
-func _SimdSse2ConditionalSquareSum(src unsafe.Pointer, srcStride, width, height uint64, mask unsafe.Pointer, maskStride uint64, value uint8, compareType uint8, sum unsafe.Pointer)
+func _SimdSse2ConditionalSquareSum(src unsafe.Pointer, srcStride, width, height uint64, mask unsafe.Pointer, maskStride uint64, value uint64/*uint8*/, compareType uint64/*uint8*/, sum unsafe.Pointer)
 
 //go:noescape
-func _SimdSse2ConditionalSquareGradientSum(src unsafe.Pointer, srcStride, width, height uint64, mask unsafe.Pointer, maskStride uint64, value uint8, compareType uint8, sum unsafe.Pointer)
+func _SimdSse2ConditionalSquareGradientSum(src unsafe.Pointer, srcStride, width, height uint64, mask unsafe.Pointer, maskStride uint64, value uint64/*uint8*/, compareType uint64/*uint8*/, sum unsafe.Pointer)
 
 //go:noescape
-func _SimdSse2ConditionalFill(src unsafe.Pointer, srcStride, width, height uint64, threshold uint8, compareType uint8, value uint8, dst unsafe.Pointer, dstStride uint64)
+func _SimdSse2ConditionalFill(src unsafe.Pointer, srcStride, width, height uint64, threshold uint64/*uint8*/, compareType uint64/*uint8*/, value uint64/*uint8*/, dst unsafe.Pointer, dstStride uint64)
 
 
 //
-func SimdSse2ConditionalCount8u(src View, value uint8, compareType uint8) uint32 {
+func SimdSse2ConditionalCount8u(src View, value uint64/*uint8*/, compareType uint64/*uint8*/) uint32 {
 
 	count := uint32(0)
 
@@ -37,7 +37,7 @@ func SimdSse2ConditionalCount8u(src View, value uint8, compareType uint8) uint32
 }
 
 //
-func SimdSse2ConditionalCount16i(src View, value int16, compareType uint8) uint32 {
+func SimdSse2ConditionalCount16i(src View, value uint64/*int16*/, compareType uint64/*uint8*/) uint32 {
 
 	count := uint32(0)
 
@@ -47,7 +47,7 @@ func SimdSse2ConditionalCount16i(src View, value int16, compareType uint8) uint3
 }
 
 //
-func SimdSse2ConditionalSum(src, mask View, value uint8, compareType uint8) uint64 {
+func SimdSse2ConditionalSum(src, mask View, value uint64/*uint8*/, compareType uint64/*uint8*/) uint64 {
 
 	sum := uint64(0)
 
@@ -57,7 +57,7 @@ func SimdSse2ConditionalSum(src, mask View, value uint8, compareType uint8) uint
 }
 
 //
-func SimdSse2ConditionalSquareSum(src, mask View, value uint8, compareType uint8) uint64 {
+func SimdSse2ConditionalSquareSum(src, mask View, value uint64/*uint8*/, compareType uint64/*uint8*/) uint64 {
 
 	sum := uint64(0)
 
@@ -67,7 +67,7 @@ func SimdSse2ConditionalSquareSum(src, mask View, value uint8, compareType uint8
 }
 
 //
-func SimdSse2ConditionalSquareGradientSum(src, mask View, value uint8, compareType uint8) uint64 {
+func SimdSse2ConditionalSquareGradientSum(src, mask View, value uint64/*uint8*/, compareType uint64/*uint8*/) uint64 {
 
 	sum := uint64(0)
 
@@ -77,7 +77,7 @@ func SimdSse2ConditionalSquareGradientSum(src, mask View, value uint8, compareTy
 }
 
 //
-func SimdSse2ConditionalFill(src View, threshold uint8, compareType uint8, value uint8, dst View) {
+func SimdSse2ConditionalFill(src View, threshold uint64/*uint8*/, compareType uint64/*uint8*/, value uint64/*uint8*/, dst View) {
 
 	_SimdSse2ConditionalFill(src.GetData(), uint64(src.GetStride()), uint64(src.GetWidth()), uint64(src.GetHeight()), threshold, compareType, value, dst.GetData(), uint64(dst.GetStride()))
 }
